requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de Datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Validación de Datos
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# Autenticación y Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Utilitarias
python-dateutil==2.8.2
pytz==2023.3.post1
